feat: enhance popup UI for history and model selection
- Introduce SetItemHl function for highlighting selected items

- Implement HistoryPreview for better history management

- Add ModelsPreview for selecting models with user feedback
