+++ lua/llm/common/api.lua
@@ -5,6 +5,7 @@ local state = require("llm.state")
 local conf = require("llm.config")
 local Popup = require("nui.popup")
 local Layout = require("nui.layout")
+local Menu = require("nui.menu")
 local LOG = require("llm.common.log")
 
 local function IsNotPopwin(winid)
@@ -644,4 +645,139 @@ function api.get_chat_ui_bufnr_list()
     return { state.llm.bufnr }
   end
 end
+
+function api.SetItemHl(popup, hl)
+  local ns = vim.api.nvim_create_namespace("SeletedItemHl")
+  local idx = vim.api.nvim_win_get_cursor(popup.winid)[1]
+  local count = vim.api.nvim_buf_line_count(popup.bufnr)
+  vim.api.nvim_buf_clear_namespace(popup.bufnr, ns, 0, count)
+
+  if vim.version.lt(vim.version(), { 0, 11, 0 }) then
+    for i = 1, count do
+      vim.api.nvim_buf_add_highlight(popup.bufnr, ns, "LlmGrayLight", i, 0, -1)
+    end
+    vim.api.nvim_buf_add_highlight(popup.bufnr, ns, hl, idx - 1, 0, -1)
+  else
+    vim.hl.range(popup.bufnr, ns, "LlmGrayLight", { 0, 0 }, { count, -1 }, {})
+    vim.hl.range(popup.bufnr, ns, hl, { idx - 1, 0 }, { idx - 1, -1 }, {})
+  end
+end
+
+function api.HistoryPreview(layout_opts, opts)
+  local layout = layout_opts or conf.configs.chat_ui_opts
+  opts = opts or conf.configs.chat_ui_opts.history.split
+  if not state.history.hl then
+    state.history.hl = opts.win_options.winhighlight:match(":(.-),")
+    opts.win_options.winhighlight = opts.win_options.winhighlight:gsub(":(.-),", ":LlmGrayLight,")
+  end
+
+  if state.history.popup == nil then
+    state.history.popup = Menu({
+      enter = opts.enter,
+      focusable = opts.focusable,
+      zindex = opts.zindex,
+      border = opts.border,
+      win_options = opts.win_options,
+      relative = opts.relative or layout.relative,
+      position = layout.position,
+      size = opts.size,
+    }, {
+      lines = (function()
+        local items = api.ListFilesInPath()
+        state.history.list = { Menu.item("current", { cmd = api.SetItemHl }) }
+        for _, item in ipairs(items) do
+          table.insert(state.history.list, Menu.item(item, { cmd = api.SetItemHl }))
+        end
+        return state.history.list
+      end)(),
+      max_width = opts.max_width,
+      keymap = {
+        focus_next = { "j", "<Down>", "<Tab>" },
+        focus_prev = { "k", "<Up>", "<S-Tab>" },
+        submit = { "<CR>", "<Space>" },
+      },
+      on_change = function(item)
+        item.cmd(state.history.popup, state.history.hl)
+        -- TODO: item index
+        if item.text == "current" then
+          state.session.filename = item.text
+          if not state.session[item.text] then
+            state.session[item.text] = api.DeepCopy(conf.session.messages)
+          end
+          api.RefreshLLMText(state.session[item.text], state.llm.bufnr, state.llm.winid, true)
+        else
+          local sess_file = string.format("%s/%s", conf.configs.history_path, item.text)
+          state.session.filename = item.text
+          if not state.session[item.text] then
+            local file = io.open(sess_file, "r")
+            if file then
+              local messages = vim.fn.json_decode(file:read())
+              state.session[item.text] = messages
+              file:close()
+            end
+          end
+          api.RefreshLLMText(state.session[item.text], state.llm.bufnr, state.llm.winid, true)
+        end
+      end,
+    })
+    state.history.popup:mount()
+    if state.history.index then
+      vim.api.nvim_win_set_cursor(state.history.popup.winid, { state.history.index, 0 })
+      local node = state.history.popup.tree:get_node(state.history.index)
+      state.history.popup._.on_change(node)
+    else
+      state.history.index = vim.api.nvim_win_get_cursor(state.history.popup.winid)[1]
+    end
+
+    state.history.popup:map("n", { "<cr>" }, function()
+      LOG:TRACE("history popup hide")
+      state.history.popup:hide()
+    end)
+
+    state.history.popup:map("n", { "<esc>" }, function()
+      local node = state.history.popup.tree:get_node(state.history.index)
+      state.history.popup._.on_change(node)
+      state.history.popup:unmount()
+      state.history.popup = nil
+    end)
+  else
+    LOG:TRACE("history popup show")
+    -- The relative winid needs to be adjusted when "relative = win",
+    if state.history.popup.border.win_config.win then
+      state.history.popup.border.win_config.win = state.llm.winid
+    end
+    state.history.popup:show()
+    state.history.index = vim.api.nvim_win_get_cursor(state.history.popup.winid)[1]
+  end
+end
+
+function api.ModelsPreview(opts, name, on_choice)
+  opts = opts or conf.configs
+  name = name or "Chat"
+  on_choice = on_choice
+    or function(choice, idx)
+      if not choice then
+        return
+      else
+        LOG:INFO("Set the current model to " .. choice)
+      end
+      opts.url, opts.model, opts.api_type, opts.max_tokens, opts.fetch_key =
+        opts.models[idx].url,
+        opts.models[idx].model,
+        opts.models[idx].api_type,
+        opts.models[idx].max_tokens,
+        opts.models[idx].fetch_key
+    end
+  state.models[name] = { list = {} }
+  for _, item in ipairs(opts.models) do
+    table.insert(state.models[name].list, item.name)
+  end
+  vim.ui.select(state.models[name].list, {
+    prompt = "Models:",
+    format_item = function(item)
+      return item
+    end,
+  }, on_choice)
+end
+
 return api
